Fixed malformed Array#reduce callbacks
Both callbacks raised an error equivalent to "Cannot set property of undefined" (the exact message varies across platforms).  Either that or the message refers to "null" instead of "undefined".

There were 2 issues:
1. The value of `o` in the callback function is always set to first element of the array if no initial value is passed to the reduce method.  Since our array elements are typically `string` or `null` and we cannot index `null`, this will lead to an error.
2. The callback function always returned `undefined` (the default when a function lacks a return statement).  Therefore, after the first loop, our callback always received `o === undefined`, which cannot be indexed.

Fixed by explicitly returning `o` in the callback and specifying an empty object (`{ }`) as the initial value of the reduce method.